Extract sync push helpers
